query menu and vendors

+++ Driver.py
@@ -55,15 +55,21 @@ def retrieve_image(session, items):
     storage_client = storage.Client()
     bucket = storage_client.get_bucket(GCP_BUCKET)
     for item in items:
-        image_name = item["image"]       
-        blob = bucket.blob(image_name)
-        tmp_image = TMP_IMAGE_PATH + image_name
-        blob.download_to_filename(tmp_image)
-        with open(tmp_image, "rb") as image_obj:
-            image_read = image_obj.read()
-            item["image_data"] = base64.b64encode(image_read)
+        image_name = item["image"]
+        print(image_name)
+        if image_name != None and image_name != "":
+            blob = bucket.blob(image_name)
+            tmp_image = TMP_IMAGE_PATH + image_name
+            blob.download_to_filename(tmp_image)
+            with open(tmp_image, "rb") as image_obj:
+                image_read = image_obj.read()
+                item["image_data"] = base64.b64encode(image_read).decode('utf-8')
+            os.remove(tmp_image)
+
+        vendor_id = item["vendor_id"]
+        vendor = session.query(Vendor).filter_by(vendor_id=vendor_id).first()
+        item["vendor_name"] = vendor.name
         info.append(item)
-        os.remove(tmp_image)
 
     return info
 
@@ -124,6 +130,7 @@ def get_menus():
         dishes = session.query(Menu).all()
         dish_info = [dish.serialize for dish in dishes]
         dish_info = retrieve_image(session, dish_info)
+        session.close()
         return jsonify(Menu = dish_info)
 
 @app.route('/menus/<int:dish_id>/')
